rewrited Makefile all

+++ libft/ft_numblen.c
@@ -6,13 +6,13 @@
 /*   By: yquaro <yquaro@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/08 14:41:41 by yquaro            #+#    #+#             */
-/*   Updated: 2019/05/01 19:04:27 by yquaro           ###   ########.fr       */
+/*   Updated: 2019/06/13 15:47:23 by yquaro           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/libft.h"
 
-int		ft_numblen(int n)
+int		ft_numblen(long long n)
 {
 	int i;
